requirements.txt
discord-py-interactions
loguru
discord.py-interactions[voice]
ffmpeg